fix auth controller conflict
