+++ routes/api/user.js
@@ -2,17 +2,11 @@ var express = require('express');
 var router = express.Router();
 
 const {
-  login,
-  register,
   getUserById,
   patchEditUser,
   deleteUser,
   getAllUser,
   uploadAvatar,
-  logout,
-  verifyEmail,
-  forgotPassword,
-  changePassword,
 } = require('../../controllers/UserController');
 
 const { avatarUploader } = require('../../middlewares/Uploader');
@@ -21,19 +15,6 @@ const { avatarUploader } = require('../../middlewares/Uploader');
 router.get('/', getAllUser);
 
 router.get('/:userId', getUserById);
-
-//METHOD POST
-router.post('/register', register);
-
-router.get('/verify-email/:userId', verifyEmail);
-
-router.post('/login', login);
-
-router.post('/logout', logout);
-
-router.post('/forgot-password', forgotPassword);
-
-router.post('/reset-password/:userId', changePassword);
 
 //upload user images
 router.post('/avatar/:userId', avatarUploader, uploadAvatar);
